add checkpoint and restore example
