+++ materials/02-Using-gem5/08-accelerating-simulation/03-checkpoint-and-restore/complete/03-take-a-checkpoint.py
@@ -1,0 +1,104 @@
+# Copyright (c) 2024 The Regents of the University of California.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met: redistributions of source code must retain the above copyright
+# notice, this list of conditions and the following disclaimer;
+# redistributions in binary form must reproduce the above copyright
+# notice, this list of conditions and the following disclaimer in the
+# documentation and/or other materials provided with the distribution;
+# neither the name of the copyright holders nor the names of its
+# contributors may be used to endorse or promote products derived from
+# this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+# OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+# SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+# LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+# DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+# THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+"""
+
+Usage:
+------
+gem5 -re --outdir=checkpointing-m5-out 03-take-a-checkpoint.py
+"""
+
+
+from gem5.components.boards.x86_board import X86Board
+from gem5.components.processors.cpu_types import CPUTypes
+from gem5.components.processors.simple_processor import SimpleProcessor
+from gem5.isas import ISA
+from gem5.resources.resource import obtain_resource
+from gem5.simulate.simulator import (
+    ExitEvent,
+    Simulator,
+)
+from gem5.utils.requires import requires
+import m5
+
+requires(
+    isa_required=ISA.X86,
+    kvm_required=True,
+)
+
+# Let's setup a NoCache cache hierarchy
+from gem5.components.cachehierarchies.classic.no_cache import NoCache
+cache_hierarchy = NoCache()
+#
+
+# Let's setup a SingleChannelDDR4_2400 memory with 3GB size
+from gem5.components.memory.single_channel import SingleChannelDDR4_2400
+memory = SingleChannelDDR4_2400(size="3GB")
+#
+
+# Here we setup a simple processor with the KVM CPU
+processor = SimpleProcessor(
+    cpu_type=CPUTypes.KVM,
+    isa=ISA.X86,
+    num_cores=2,
+)
+#
+
+for proc in processor.get_cores():
+    proc.core.usePerf = False
+
+board = X86Board(
+    clk_freq="3GHz",
+    processor=processor,
+    memory=memory,
+    cache_hierarchy=cache_hierarchy,
+)
+
+board.set_workload(obtain_resource("npb-ep-a"))
+
+# Setup workbegin handler to reset stats and switch to TIMING CPU
+def workbegin_handler():
+    print("Done booting Linux")
+
+    print("Take a checkpoint")
+    simulator.save_checkpoint("03-cpt")
+
+    yield True
+#
+
+simulator = Simulator(
+    board=board,
+    on_exit_event= {
+        ExitEvent.WORKBEGIN: workbegin_handler(),
+    }
+)
+
+print("Running the simulation")
+print("Using KVM cpu")
+
+simulator.run()
+
+print("Simulation Done")
